moved functionality in ReflectUtil to ProtoUtil and made it private
